+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Allow selection of multiple degrees to have all the courses in the selection list
+
 ### Fixed
 
 - Add temporary favicon.ico